http接口有变
地点和车辆状态接口有变

目前三个模块可以正常运行，简单数据测试有效

+++ fleet-schedule-demo/data-cache/src/main/java/demo/Service/RoadService.java
@@ -13,8 +13,8 @@ public interface RoadService {
     List<Road> saveRoads(List<Road> roads);
     void deleteAll();
     List<Road> findAll();
-    Page<Road> findByBeginPoint(Long beginPoint, Pageable pageable);
-    Page<Road> findByEndPoint(Long endPoint, Pageable pageable);
-    Page<Road> findByBeginPointAndEndPoint(Long beginPoint, Long endPoint, Pageable pageable);
+    Page<Road> findByBeginPoint(Integer beginPoint, Pageable pageable);
+    Page<Road> findByEndPoint(Integer endPoint, Pageable pageable);
+    Page<Road> findByBeginPointAndEndPoint(Integer beginPoint, Integer endPoint, Pageable pageable);
 
 }
