+++ fleet-schedule-demo/data-cache/src/main/java/demo/Service/impl/RoadServiceImpl.java
@@ -36,17 +36,17 @@ public class RoadServiceImpl implements RoadService {
     }
 
     @Override
-    public Page<Road> findByBeginPoint(Long beginPoint, Pageable pageable) {
+    public Page<Road> findByBeginPoint(Integer beginPoint, Pageable pageable) {
         return this.roadRepository.findByBeginPoint(beginPoint, pageable);
     }
 
     @Override
-    public Page<Road> findByEndPoint(Long endPoint, Pageable pageable) {
+    public Page<Road> findByEndPoint(Integer endPoint, Pageable pageable) {
         return this.roadRepository.findByEndPoint(endPoint, pageable);
     }
 
     @Override
-    public Page<Road> findByBeginPointAndEndPoint(Long beginPoint, Long endPoint, Pageable pageable) {
+    public Page<Road> findByBeginPointAndEndPoint(Integer beginPoint, Integer endPoint, Pageable pageable) {
         return this.roadRepository.findByBeginPointAndEndPoint(beginPoint, endPoint, pageable);
     }
 }
