+++ fleet-schedule-demo/data-cache/src/main/java/demo/domain/Road.java
@@ -3,6 +3,7 @@ package demo.domain;
 //道路数据
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.sun.org.apache.xalan.internal.xsltc.util.IntegerArray;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -23,8 +24,8 @@ public class Road {
     @Id
     private Long rid;  //道路编号
 
-    private Long beginPoint; //起点编号
-    private Long endPoint;  //终点编号
+    private Integer beginPoint; //起点编号
+    private Integer endPoint;  //终点编号
     private double distance;  //道路长度
     //private double curNumPoint; //加入该道路后总地点数
